+++ pkg/kubectl/cmd/get/humanreadable_flags.go
@@ -21,7 +21,7 @@ import (
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/kubernetes/pkg/printers"
+	"k8s.io/cli-runtime/pkg/printers"
 )
 
 // HumanPrintFlags provides default flags necessary for printing.
